+++ src/main/java/kr/ac/skuniv/realestate/domain/entity/RentDate.java
@@ -22,10 +22,10 @@ public class RentDate {
     private Date date;
 
     @Column(name = "guarantee_price")
-    private Double guaranteePrice;
+    private Integer guaranteePrice;
 
     @Column(name = "monthly_price")
-    private Double monthlyPrice;
+    private Integer monthlyPrice;
 
     @ManyToOne(cascade = {CascadeType.ALL}, fetch = FetchType.LAZY)
     @JoinColumn(name = "building_id", insertable = false, updatable = false)
